fix Driver::getMotorState return type

+++ src/Components/Driver.cpp
@@ -40,7 +40,7 @@ void Driver::setMotor(uint8_t id, int8_t state){
 	Motors.setMotor(id, state);
 }
 
-uint Driver::getMotorState(int id){
+int8_t Driver::getMotorState(uint8_t id){
 	return motors[id];
 }
